# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TraclyTag Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/stock:
    get:
      operationId: getStockReport
      tags: [reports]
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StockRow"
  /reports/product:
    get:
      operationId: getProductReport
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductReportRow"
  /reports/shipper-summary:
    get:
      operationId: getShipperSummary
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShipperSummaryRow"
  /reports/pallet-summary:
    get:
      operationId: getPalletSummary
      tags: [reports]
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PalletSummaryRow"
...
components:
  schemas:
...
    StockRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        totalCodes:
          type: integer
        mapped:
          type: integer
        unmapped:
          type: integer
        location:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - totalCodes
        - mapped
        - unmapped
        - location
    ProductReportRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        size:
          type: string
        total:
          type: integer
        mapped:
          type: integer
        unmapped:
          type: integer
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - size
        - total
        - mapped
        - unmapped
    ShipperSummaryRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        size:
          type: string
        total:
          type: integer
        shipperMapped:
          type: integer
        shipperUnmapped:
          type: integer
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - size
        - total
        - shipperMapped
        - shipperUnmapped
    PalletSummaryRow:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        batchId:
          type: integer
        batchNumber:
          type: string
        size:
          type: string
        total:
          type: integer
        palletMapped:
          type: integer
        palletUnmapped:
          type: integer
      required:
        - productId
        - productName
        - batchId
        - batchNumber
        - size
        - total
        - palletMapped
        - palletUnmapped
...